refact: organize redis connector on src
